Fix multi-disk bug
On systems with large amounts of disks (more than a single-digit), it would be possible for grep to prefix-match multiple disks.
For example searching for "disk1" would return all disks starting with 1, i.e. "disk10", "disk11" and so on.
This fix adds an "s" after the input string to the findEFIPartition method, which disables this behavior as now matches only occur for the full disk number instead of just parts of it.